QuickTime/MP4 reader: JSON ident: report audio sampling frequency as an integer
My JSON schema says it should be an integer, so make it one and not a
floating point number.

All other input modules already either use integers for storing the
value or cast the floating point variables to integers before adding
them to the JSON output.

This will only required required with an updated version of nlohmann's
JSON library as the current library implicitly casts the float to an
int. The updated version won't do that anymore and preserve the type.

+++ src/input/r_qtmp4.cpp
@@ -1846,7 +1846,7 @@ qtmp4_reader_c::identify() {
 
     else if (dmx->is_audio()) {
       info.add(mtx::id::audio_channels,           dmx->a_channels);
-      info.add(mtx::id::audio_sampling_frequency, dmx->a_samplerate);
+      info.add(mtx::id::audio_sampling_frequency, static_cast<uint64_t>(dmx->a_samplerate));
       info.add(mtx::id::audio_bits_per_sample,    dmx->a_bitdepth);
     }
